Optimized some things. Added and updated diagrams.

+++ VegiJ.Logic/RecipeManager.cs
@@ -3,7 +3,7 @@
     using System;
     using System.Linq;
     using VegiJ.DataAccess.Contracts;
-
+    // TODO: Implement later reporting for dublicate recipes? or nah
     public class RecipeManager : IRecipeManager
     {
         private IRepository<Recipe> _recipeRepository;
@@ -13,14 +13,24 @@
             this._recipeRepository = recipeRepository;
         }
 
-        // TODO: Add recipe exist method
+        // TODO: Improve recipe exist method (maybe with regex, etc.)
         public void AddRecipe(Recipe recipe)
         {
+            if (CheckIfRecipeExist(recipe))
+            {
+                return;
+            }
+
             this._recipeRepository.Create(recipe);
         }
 
         public void DeleteRecipe(Recipe recipe)
         {
+            if (!CheckIfRecipeExist(recipe))
+            {
+                return;
+            }
+
             this._recipeRepository.Delete(recipe.ID);
         }
 
@@ -36,7 +46,20 @@
 
         public void UpdateRecipe(Recipe recipe)
         {
+            if (CheckIfRecipeExist(recipe))
+            {
+                return;
+            }
+
             this._recipeRepository.Update(recipe);
+        }
+
+        private bool CheckIfRecipeExist(Recipe recipe)
+        {
+            var thereIsSameTitle = this._recipeRepository.Table.Any(r => r.Title.Equals(recipe.Title, StringComparison.InvariantCultureIgnoreCase));
+            var thereIsSameContent = this._recipeRepository.Table.Any(r => r.Content.Equals(recipe.Content, StringComparison.InvariantCultureIgnoreCase));
+
+            return thereIsSameContent || thereIsSameTitle;
         }
     }
 }
